Refactor contents proxy into separate file
